issue-521: updated snmp testcases
